+++ docs/changelog.rst
@@ -10,7 +10,11 @@ Version 0.8.0 - Unreleased
 * Relaxed ``libsass`` version in requirements;
 * Moved ``colorama`` from test to default requirements;
 * Removed every use of click.echo/secho within core API, use logger instead, close #1;
-* Added ``colorlog`` in requirements and use it to have colors for each logging level #4;
+* Added ``colorlog`` in requirements and use it to have colors for each logging level, close #4;
+
+* Changed verbosity option on CLI action so default verbosity is INFO logging level, then user can choose totally silent verbosity or any other logging level, definitively close #1;
+* Better CLI actions helps, close #5;
+* Manage every API exception from CLI, should be ok now (in fact since previous commit), close #3;
 
 Version 0.7.0 - 2016/04/07
 --------------------------